feat: 调整接口

+++ src/module/staff/staff.controller.ts
@@ -2,7 +2,7 @@ import { Controller, Post, Body } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { NoLogin } from '../../common/decorator/auth.decorator';
 import ErrorCode from '../../constant/error-code';
-import { resSuccess, Utils } from '../../utils/index';
+import { resSuccess } from '../../utils/index';
 import { HttpResponse, ListResponse } from '../../types/type';
 import { StaffService } from './staff.service';
 import {
@@ -22,7 +22,7 @@ export class StaffController {
   ) {}
 
   /*
-   * 查询全部用户
+   * 查询全部账号
    */
   @Post('get-list')
   @NoLogin
@@ -62,7 +62,7 @@ export class StaffController {
     if (!staff) {
       return {
         code: ErrorCode.STAFF_NOT_FOUND,
-        message: '用户不存在',
+        message: '账号不存在',
       };
     }
 
@@ -75,7 +75,7 @@ export class StaffController {
   }
 
   /*
-   * 新增用户
+   * 新增账号
    */
   @Post('add')
   @NoLogin
@@ -84,39 +84,48 @@ export class StaffController {
     const sameNameStaff = await this.StaffService.getDetailByName(createStaffDto.name);
     if (sameNameStaff) {
       return {
-        code: ErrorCode.STAFF_USERNAME_UNIQUE,
-        message: '用户名已存在',
+        code: ErrorCode.STAFF_NAME_UNIQUE,
+        message: '账号名已存在',
       };
     }
 
-    // 生成随机的staffId
-    const staffId = Utils.generateId('staff');
-    const staff = {
-      ...createStaffDto,
-      staffId,
-    };
-
     // 写入数据库
-    const res = this.StaffService.create(staff);
+    const res = this.StaffService.create(createStaffDto);
     return resSuccess(res);
   }
 
   /*
-   * 更新用户
+   * 更新账号
    */
   @Post('update')
   @NoLogin
   async update(@Body() updateStaffDto: UpdateStaffDto): Promise<HttpResponse<any>> {
+    const staff = await this.StaffService.getDetail(updateStaffDto.staffId);
+    if (!staff) {
+      return {
+        code: ErrorCode.STAFF_NOT_FOUND,
+        message: '账号不存在',
+      };
+    }
+
     const arr = this.StaffService.update(updateStaffDto);
     return resSuccess(arr);
   }
 
   /*
-   * 删除用户
+   * 删除账号
    */
   @Post('delete')
   @NoLogin
   async delete(@Body() deleteStaffDto: DeleteStaffDto): Promise<HttpResponse<any>> {
+    const staff = await this.StaffService.getDetail(deleteStaffDto.staffId);
+    if (!staff) {
+      return {
+        code: ErrorCode.STAFF_NOT_FOUND,
+        message: '账号不存在',
+      };
+    }
+
     const res = await this.StaffService.delete(deleteStaffDto.staffId);
     return resSuccess(res);
   }
